# docs/swagger.yaml
definitions:
  device.Coordinates:
    properties:
      latitude:
        default: 37.552375
        type: number
      longitude:
        default: 55.646575
        type: number
    type: object
  device.DeviceRequest:
    properties:
      coordinates:
        $ref: '#/definitions/device.Coordinates'
      email:
        default: some@email.com
        type: string
      language:
        default: en-US
        type: string
      platform:
        default: macOs
        type: string
      uuid:
        default: f2366ac9-0663-4d0b-964f-98c388240d5c
        type: string
    type: object
  device.UpdateEmailRequest:
    properties:
      email:
        type: string
      uuid:
        type: string
    type: object
  device.UpdateGeolocationRequest:
    properties:
      latitude:
        default: 37.552375
        type: number
      longitude:
        default: 55.646575
        type: number
      uuid:
        type: string
    type: object
  device.UpdateLanguageRequest:
    properties:
      language:
        type: string
      uuid:
        type: string
    type: object
  domain.Device:
    properties:
      email:
        type: string
      language:
        type: string
      location:
        $ref: '#/definitions/domain.Location'
      platform:
        type: string
      uuid:
        type: string
    type: object
  domain.Event:
    properties:
      attributes:
        items: {}
        type: array
      createdAt:
        type: string
      deviceUUID:
        type: string
      name:
        type: string
    type: object
  domain.Location:
    properties:
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
  event.addEventRequest:
    properties:
      attributes:
        items: {}
        type: array
      name:
        type: string
      uuid:
        type: string
    type: object
  handlers.ErrorResponce:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: string
    type: object
info:
  contact: {}
  title: Device Manager API
  version: "1.0"
paths:
  /device:
    post:
      consumes:
      - application/json
      description: Add a new device to database
      parameters:
      - description: Add Device Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/device.DeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Add device
      tags:
      - Device
  /device/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete a device from database
      parameters:
      - description: Device Id
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Delete a device
      tags:
      - Device
    get:
      consumes:
      - application/json
      description: Get device info
      parameters:
      - description: Device UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "404":
          description: Not Found
      summary: Get a device
      tags:
      - Device GET
  /device_email:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Device E-mail Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/device.UpdateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.UpdateEmailRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Update a device E-mail
      tags:
      - Device UPDATE
  /device_email/{email}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Devices Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Device'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Get devices by Email filter
      tags:
      - Device GET
  /device_geo:
    get:
      consumes:
      - application/json
      parameters:
      - description: longitude
        in: query
        name: longitude
        required: true
        type: number
      - description: latitude
        in: query
        name: latitude
        required: true
        type: number
      - description: distance
        in: query
        name: distance
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Get devices by Geoposition
      tags:
      - Device GET
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Device geolocation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/device.UpdateGeolocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.UpdateGeolocationRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Update a device geolocation
      tags:
      - Device UPDATE
  /device_lang:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Device language Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/device.UpdateLanguageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/device.UpdateLanguageRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Update a device language
      tags:
      - Device UPDATE
  /device_lang/{language}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Devices Language
        in: path
        name: language
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Device'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Get devices by Language filter
      tags:
      - Device GET
  /event:
    get:
      consumes:
      - application/json
      description: Get events of device from database
      parameters:
      - description: UUID
        in: query
        name: uuid
        required: true
        type: string
      - description: Begin time range
        in: query
        name: timeBegin
        required: true
        type: string
      - description: End time range
        in: query
        name: timeEnd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Get events
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: Add a new event from device to database
      parameters:
      - description: Add Event Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.addEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponce'
      summary: Add event
      tags:
      - Event
swagger: "2.0"